cleanup types and refactor context
