+++ frontend/src/context/filter/filter-context.tsx
@@ -2,34 +2,12 @@
 
 // frontend/src/context/jobs/filter-context.tsx
 import { createContext, useContext } from "react";
-import { JobFilters, SortBy } from "@/types/filters";
-
-export interface FilterState {
-  filters: JobFilters;
-  totalJobs: number;
-  isLoading: boolean;
-  error: Error | null;
-}
+import { FilterState } from "@/types/filters";
 
 interface FilterContextType {
   filters: FilterState;
   updateFilters: (filters: Partial<FilterState>) => void;
 }
-
-export const initialState: FilterState = {
-  filters: {
-    search: "",
-    industryFields: [],
-    jobTypes: [],
-    locations: [],
-    workingRights: [],
-    page: 1,
-    sortBy: SortBy.RECENT,
-  },
-  totalJobs: 0,
-  isLoading: false,
-  error: null,
-};
 
 export const FilterContext = createContext<FilterContextType | undefined>(
   undefined,
