+++ frontend/src/types/job.ts
@@ -1,33 +1,42 @@
 // frontend/src/types/job.ts
-export type JobType = "EOI" | "FIRST_YEAR" | "INTERN" | "GRADUATE" | "OTHER";
-export type LocationType =
-  | "VIC"
-  | "NSW"
-  | "QLD"
-  | "WA"
-  | "NT"
-  | "SA"
-  | "ACT"
-  | "TAS"
-  | "AUSTRALIA"
-  | "OTHERS";
-export type WFHStatus = "HYBRID" | "REMOTE" | "OFFICE";
-export type WorkingRight =
-  | "AUS_CITIZEN"
-  | "AUS_PR"
-  | "NZ_CITIZEN"
-  | "NZ_PR"
-  | "INTERNATIONAL"
-  | "WORK_VISA"
-  | "VISA_SPONSORED"
-  | "OTHER_RIGHTS";
-export type IndustryField =
-  | "CONSULTING"
-  | "BANKS"
-  | "BIG_TECH"
-  | "TECH"
-  | "QUANT_TRADING"
-  | "OTHER_INDUSTRY";
+export const JOB_TYPES = ["FIRST_YEAR", "INTERN", "GRADUATE", "OTHER"] as const;
+export type JobType = (typeof JOB_TYPES)[number];
+
+export const LOCATIONS = [
+  "VIC",
+  "NSW",
+  "QLD",
+  "WA",
+  "NT",
+  "SA",
+  "ACT",
+  "TAS",
+  "AUSTRALIA",
+  "OTHERS",
+] as const;
+export type LocationType = (typeof LOCATIONS)[number];
+
+export const WORKING_RIGHTS = [
+  "AUS_CITIZEN",
+  "AUS_PR",
+  "NZ_CITIZEN",
+  "NZ_PR",
+  "INTERNATIONAL",
+  "WORK_VISA",
+  "VISA_SPONSORED",
+  "OTHER_RIGHTS",
+] as const;
+export type WorkingRight = (typeof WORKING_RIGHTS)[number];
+
+export const INDUSTRY_FIELDS = [
+  "CONSULTING",
+  "BANKS",
+  "BIG_TECH",
+  "TECH",
+  "QUANT_TRADING",
+  "OTHER_INDUSTRY",
+] as const;
+export type IndustryField = (typeof INDUSTRY_FIELDS)[number];
 
 export interface Company {
   name: string;
@@ -41,7 +50,6 @@ export interface Job {
   description?: string;
   company: Company;
   applicationUrl?: string;
-  wfhStatus?: WFHStatus;
   sourceUrls: string[];
   type?: JobType;
   openDate?: string;
